requirements.txt
aenum==2.2.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
docopt==0.6.2
eventlet==0.25.1
Flask==1.1.1
Flask-SocketIO==4.2.1
greenlet==0.4.15
hx711==1.1.2.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.18.2
pipreqs==0.4.10
pkg-resources==0.0.0
psutil==5.7.0
python-can==3.3.2
python-engineio==3.12.1
python-socketio==4.5.1
redis==3.4.1
requests==2.23.0
RPi.GPIO==0.7.0
six==1.14.0
smbus==1.1.post2
tinkerforge==2.1.25
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.0
wrapt==1.12.1
yarg==0.1.9
